fixed most initial tests
